Release v0.3.0: production hardening, compression fixes, Python 3.12+ (#17)
* fix: harden decoding and surface server errors

* fix: preserve streamed blocks and connection state

* fix: surface sync server exceptions

* fix: enforce Python pool exclusivity

* fix: isolate per-query Python settings

* fix: enforce connection setup timeouts

* fix: unify query ID generation

* fix: cap native framing allocations

* test: gate async query ID regression

* fix: cap native protocol item counts

* fix: cap column byte allocations

* docs: note QueryStream chunk cap

* fix: make async cancellation safe

* fix: make Python cancellation kill queries

* fix: frame buffered nested columns correctly

* fix: enforce cumulative response budgets

* ci: repair nextest profile and release gates

* chore: trim packaging and fix docs

* fix: stop Python pool acquire starvation

* fix: update pyo3 and widen Python support

* fix: harden plain column type layout

* fix: drain consumed stream bytes

* test: add column fuzz harnesses

* perf: batch wire reads and sweep pool slots

* test: gate part UUID tests by server version

* test: stress free-threaded shared state

* fix: decompress multi-frame responses

* fix: cap exception chain depth

* docs: align APIs and fix examples

* docs: add v0.3 release readiness report

* chore: release v0.3.0

* docs: polish release notes and add v0.4 backlog

* ci: make crate publishing idempotent

* ci: release on merge with version gate

* docs: fix diagram and use percentages

* docs: clarify benchmarks and coverage

* docs: sharpen crate synopsis

* docs: restructure quick starts and refresh python numbers

* chore: keep docs local and add versioning to contributing

+++ src/sync/column/any.rs
@@ -89,6 +89,9 @@ impl<'a> AnyColumnData<'a> {
     /// string-like values use `String`.
     pub fn into_owned(&self) -> OwnedColumnData {
         let len = self.len();
+        // `unwrap_or(0)` below is unreachable-by-construction: the loop is
+        // bounded by `len == col.len()`, and `get` only errors on
+        // out-of-bounds indices.
         match self {
             Self::UInt8(col) => {
                 let mut v = Vec::with_capacity(len);
@@ -579,7 +582,7 @@ pub fn read_column_by_type<'a>(
             }))
         },
         LowCardinality(inner) => read_column_by_type(inner, ctx),
-        Date32 => u16::read_column(ctx).map(AnyColumnData::UInt16),
+        Date32 => i32::read_column(ctx).map(AnyColumnData::Int32),
         Nothing => {
             let _ = ctx.read_exact(ctx.rows)?;
             Ok(AnyColumnData::Unknown)
